move inside components
