+++ src/app/[orgId]/settings/sites/[niceId]/configuration/siteConfigComponents/resourceItem.tsx
